move location related files out of lens lib folder
