+++ apps/app/src/components/Causes/CauseCard.tsx
@@ -4,8 +4,8 @@ import { toast } from 'react-hot-toast'
 import { useTranslation } from 'react-i18next'
 
 import { CURRENCIES } from '@/constants'
+import { formatLocation } from '@/lib/formatLocation'
 import getAvatar from '@/lib/getAvatar'
-import { formatLocation } from '@/lib/lens-protocol/formatLocation'
 import lensClient from '@/lib/lens-protocol/lensClient'
 import { CauseMetadata, isComment, isPost, PostTags } from '@/lib/metadata'
