+++ apps/app/src/lib/lens-protocol/index.ts
@@ -1,7 +1,6 @@
 import checkAuth from './checkAuth'
 import createCollect from './createCollect'
 import createProfile from './createProfile'
-import { formatLocation } from './formatLocation'
 import getProfile from './getProfile'
 import getProfilesOwnedBy from './getProfilesOwnedBy'
 import getSignature from './getSignature'
@@ -13,7 +12,6 @@ import useCreatePost from './useCreatePost'
 import useEnabledCurrencies from './useEnabledCurrencies'
 import useExplorePublications from './useExplorePublications'
 import useFollow from './useFollow'
-import useLazyGeolocation from './useLazyGeolocation'
 import usePostData from './usePostData'
 import usePublication from './usePublication'
 
@@ -21,7 +19,6 @@ export {
   checkAuth,
   createCollect,
   createProfile,
-  formatLocation,
   getProfile,
   getProfilesOwnedBy,
   getSignature,
@@ -33,7 +30,6 @@ export {
   useEnabledCurrencies,
   useExplorePublications,
   useFollow,
-  useLazyGeolocation,
   usePostData,
   usePublication
 }
